# docs/swagger.yaml
basePath: /
definitions:
  github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.MessageResponse:
    properties:
      message:
        type: string
    type: object
  ? github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.PaginatedResponse-github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark_BookmarkInfo
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark.BookmarkInfo'
        type: array
      pagination:
        $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.Pagination'
    type: object
  github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  ? github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.Response-github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark_BookmarkInfo
  : properties:
      data:
        $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark.BookmarkInfo'
      message:
        type: string
    type: object
  github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark.BookmarkInfo:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark.NewBookmarkRequest:
    properties:
      description:
        type: string
      url:
        type: string
    required:
    - url
    type: object
  github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark.UpdateBookmarkRequest:
    properties:
      description:
        type: string
      url:
        type: string
    required:
    - url
    type: object
  internal_handler_health_check.HealthResponse:
    properties:
      instance_id:
        type: string
      message:
        type: string
      service_name:
        type: string
    type: object
  internal_handler_shorten.shortenInputBody:
    properties:
      exp:
        maximum: 604800
        type: integer
      url:
        type: string
    required:
    - exp
    - url
    type: object
  internal_handler_shorten.shortenResMessage:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: This is a book API
  title: Book API
  version: "1.0"
paths:
  /health-check:
    get:
      description: Check the health of the API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_health_check.HealthResponse'
      summary: Health check endpoint
      tags:
      - health_check
  /v1/bookmarks:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.PaginatedResponse-github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark_BookmarkInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.MessageResponse'
      security:
      - BearerAuth: []
      summary: Get bookmarks
      tags:
      - bookmark
    post:
      consumes:
      - application/json
      parameters:
      - description: Input required
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark.NewBookmarkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.Response-github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark_BookmarkInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.Response-github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark_BookmarkInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.Response-github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark_BookmarkInfo'
      security:
      - BearerAuth: []
      summary: Create bookmark
      tags:
      - bookmark
  /v1/bookmarks/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bookmark id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.MessageResponse'
      security:
      - BearerAuth: []
      summary: Delete bookmark
      tags:
      - bookmark
    put:
      consumes:
      - application/json
      parameters:
      - description: Bookmark id
        in: path
        name: id
        required: true
        type: string
      - description: Input required
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api_bookmark.UpdateBookmarkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_bookmark-project-learn_bookmark-service_internal_models_dto_api.MessageResponse'
      security:
      - BearerAuth: []
      summary: Update bookmark
      tags:
      - bookmark
  /v1/links/redirect/{code}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Shorten code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
      tags:
      - link
  /v1/links/shorten:
    post:
      consumes:
      - application/json
      description: Generate shorten url based on original url that last upto 7 days
      parameters:
      - description: Input required
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handler_shorten.shortenInputBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_shorten.shortenResMessage'
      summary: Generate shorten url based on original url that last upto 7 days
      tags:
      - link
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"